rename page to use the final route
