+++ pages/pricing/index.tsx
@@ -1,8 +1,14 @@
 import { NextPage } from 'next'
-import Image from 'next/legacy/image'
-import Head from 'next/head'
-import { useCallback, useEffect, useState } from 'react'
-import CheckMark from '../../public/images/checkmark.svg'
+import Image from 'next/image'
+import { PrimaryButton } from '../../components/common/Buttons/PrimaryButton'
+import { FooterCallToAction } from '../../components/common/CallToAction/FooterCallToAction'
+import Footer from '../../components/common/Footer/Footer'
+import Navbar from '../../components/common/Navbar/Navbar'
+import { Typography } from '../../components/common/Typography/Typography'
+import homeStyles from '../../components/Home/Home.module.scss'
+import pricingStyles from '../../components/Pricing/Pricing.module.scss'
+import { InformationCircleIcon, ChevronDownIcon, CheckIcon } from '@heroicons/react/20/solid'
+
 import PcPlayMedia from '../../public/images/pc-play-media.svg'
 import Wallet from '../../public/images/wallet.svg'
 import Stopwatch from '../../public/images/stopwatch.svg'
@@ -12,561 +18,506 @@ import ReceiptList from '../../public/images/receipt-list.svg'
 import CreditCard from '../../public/images/credit-card.svg'
 import Delete from '../../public/images/delete.svg'
 import TagLoyalty from '../../public/images/tag-loyalty.svg'
-import Chat from '../../public/images/pricing-comment.svg'
-import Navbar from '../../components/common/Navbar/Navbar'
-import { Section } from '../../components/common/Section/Section'
-import Footer from '../../components/common/Footer/Footer'
-import styles from '../../components/Pricing/Pricing.module.scss'
-import productStyles from '../../components/Products/Products.module.scss'
+
+import { RadioGroup, Listbox } from '@headlessui/react'
+import * as Slider from '@radix-ui/react-slider'
+import { useState } from 'react'
 import classNames from 'classnames'
-import { PrimaryButton } from '../../components/common/Buttons/PrimaryButton'
-import { FooterCallToAction } from '../../components/common/CallToAction/FooterCallToAction'
-import { Typography } from '../../components/common/Typography/Typography'
 import { CompaniesReel } from '../../components/Home/CompaniesReel/CompaniesReel'
-import useEmblaCarousel from 'embla-carousel-react'
-
 import Collapsible from 'react-collapsible'
-import {
-  BasicDetails,
-  EnterpriseDetails,
-  EssentialsDetails,
-  PricingDetails,
-  StartupDetails,
-} from '../../components/Pricing/plan_features'
-import { BasicInfo, EnterpriseInfo, EssentialsInfo, PricingInfo, StartupInfo } from '../../components/Pricing/plan_info'
-import { Meta } from '../../components/common/Head/Meta'
-import { useMediaQuery } from '../../components/MediaQuery/MediaQuery'
+import { Section } from '../../components/common/Section/Section'
 
-// Plans and info used for both mobile and desktop views.
-const planDetails: Array<PricingDetails> = [BasicDetails, EssentialsDetails, StartupDetails, EnterpriseDetails]
-const planInfo: Array<PricingInfo> = [BasicInfo, EssentialsInfo, StartupInfo, EnterpriseInfo]
-
-const docsUrl = '/docs'
-
-const FAQS = [
-  {
-    question: 'Do you offer a discount for non-profits?',
-    answer: `We love supporting non-profits and offer a 75% discount for the lifetime of the account. To activate the discount, create a workplace on either the Standard or Pro plan. Then reach out to support and mention the discount. `,
-    icon: TagLoyalty,
-  },
-  {
-    question: 'How long does it take to setup Highlight?',
-    answer: `It generally takes an engineer less than ten minutes to understand the concepts of Highlight and integrate the app into their workflow. For more information on setup, take a look at our <a href="${docsUrl}">docs</a>. `,
-    icon: Stopwatch,
-  },
-  {
-    question: 'Can I deploy Highlight on-premise?',
-    answer: `Yes! To get a glimpse at how our deployment process looks, take a look <a href="${docsUrl}/on-premise">here</a> (its super simple!). We also support deployments for most cloud providers (Heroku, Render, AWS, etc..). To get a license key for a trial or a production deployment, contact <a href="mailto:sales@highlight.io">sales</a>. `,
-    icon: Globe,
-  },
-  {
-    question: "Is Highlight secure? Where's my data stored?",
-    answer: `Highlight uses end-to-end encryption to keep your data safe while it’s in transit, and we also offer an on-prem solution if you want to keep customer data on your own servers. For more information, see our <a href="/#privacy">security section</a> and <a href="${docsUrl}" target="_blank">docs</a>. If we don't answer your question there, <a href="mailto:jay@highlight.io">let us know</a>. `,
-    icon: Security,
-  },
-  {
-    question: 'Do I need a credit card to sign up?',
-    answer: `Absolutely not! We never ask for your credit card on sign up. If you start on a paid plan then 30 days after signing up you will be politely prompted to enter in your payment information. At anytime you can switch back to a free plan as long as your workplace has less than 6 seats. `,
-    icon: CreditCard,
-  },
-  {
-    question: 'How will you charge me?',
-    answer: `We ask for a credit card. Your credit card information will never touch our servers as we use <a href="https://stripe.com/" target="_blank">Stripe</a> as our payments processor. For Enterprise customers we can do ACH and custom invoices if requested. `,
-    icon: Wallet,
-  },
-  {
-    question: 'How does billing work?',
-    answer: `We charge by usage; or number of sessions collected per month. Our billing system uses prorated billing, meaning you only pay for what you use below each of our thresholds (see above). For example if you move to the Startup plan from the Basic plan in the middle of the month, then you will only be charged for the time you are on the paid plan. `,
-    icon: ReceiptList,
-  },
-  {
-    question: 'What counts as a session?',
-    answer: `A session is contiguous instance of a user's presence on your app for less than 4 hours. That is, if a user is browsing your application for 3 minutes, then closes the tab, this counts as a single session. `,
-    icon: PcPlayMedia,
-  },
-  {
-    question: 'Can I cancel at anytime?',
-    answer: `Definitely! You can cancel or downgrade your subscription at anytime. You can also delete your workplace in the settings page at anytime. `,
-    icon: Delete,
-  },
-]
-
-const TierSection = ({
-  tierName,
-  numSessionCredits,
-  price,
-  discount,
-  features,
-  mostPopular,
-  monthly,
-}: PricingInfo & {
-  features: PricingDetails
-  monthly: boolean
-}) => {
-  const contactSales = price === undefined
+const PricingPage: NextPage = () => {
   return (
     <div>
-      {mostPopular && (
-        <div className={styles.mostPopular}>
-          <Typography type="outline">Most Popular</Typography>
-        </div>
-      )}
-      <div
-        className={classNames(styles.tierColumn, {
-          [styles.mostPopularTierColumn]: mostPopular,
-        })}
-      >
-        <div className={styles.desktopTierSection}>
-          <div className={styles.desktopTopTier}>
-            <div className={styles.tierHeader}>
-              <Typography type="copy1" emphasis>
-                {tierName}
-              </Typography>
-              <div
-                className={classNames(styles.discountPill, {
-                  [styles.discountPillVisible]: !monthly && discount,
-                })}
-              >
-                - {discount}%
-              </div>
-            </div>
-            {contactSales ? (
-              <div className={styles.desktopSessionCreditsEnterprise}>
-                <Typography type="copy3" emphasis>
-                  {'Custom'}
-                </Typography>
-                <Typography type="copy3">{` session credits`}</Typography>
-              </div>
-            ) : (
-              <div className={styles.desktopSessionCredits}>
-                <Typography type="copy3" emphasis>
-                  {numSessionCredits.toLocaleString()}
-                </Typography>
-                <Typography type="copy3">{` session credits`}</Typography>
-              </div>
-            )}
-            <div className={styles.desktopPrice}>
-              {contactSales ? (
-                <Image height={24} width={24} src={Chat} alt="chat icon"></Image>
-              ) : (
-                <>
-                  <Typography type="copy3" emphasis className={styles.moneySign}>
-                    {'$'}
-                  </Typography>
-                  <h3 className={styles.price}>
-                    {monthly || !discount ? price : Math.round((1 - discount / 100) * price)}
-                  </h3>
-                  <div className={styles.timeIndicator}>
-                    <Typography type="copy3">{'/ mo'}</Typography>
-                  </div>
-                </>
-              )}
-            </div>
-          </div>
-          <PrimaryButton
-            href={contactSales ? 'mailto:sales@highlight.io' : 'https://app.highlight.io/?sign_up=1'}
-            className={styles.pricingButton}
-          >
-            <Typography type="copy2" emphasis={true}>
-              {contactSales ? 'Contact Sales' : 'Start Free Trial'}
-            </Typography>
-          </PrimaryButton>
-        </div>
-
-        {features &&
-          Object.keys(features).map((headingKey) => {
-            return (
-              <div className={styles.desktopFeaturesSection} key={headingKey}>
-                {(features as any)[headingKey].items.map((item: any) => {
-                  return (
-                    <>
-                      {item.value ? <Image src={CheckMark} alt="checkmark"></Image> : <div>-</div>}
-                      <hr className={styles.featureDivider} />
-                    </>
-                  )
-                })}
-              </div>
-            )
-          })}
-      </div>
-    </div>
-  )
-}
-
-const Pricing: NextPage = () => {
-  const [monthly, setMonthly] = useState(true)
-  const is800 = useMediaQuery(800)
-  const is400 = useMediaQuery(400)
-  const isMobile = is800 || is400
-  return (
-    <div>
-      <Head>
-        <title>Highlight: Plans And Pricing For Any Team. Get Started Free.</title>
-        <script type="application/ld+json">
-          {`{
-            "@context": "https://schema.org",
-            "@type": "FAQPage",
-            "mainEntity":
-              ${JSON.stringify(
-                FAQS.map((faq) => ({
-                  '@type': 'Question',
-                  name: faq.question,
-                  acceptedAnswer: {
-                    '@type': 'Answer',
-                    text: faq.answer,
-                  },
-                })),
-              )}
-          }`}
-        </script>
-      </Head>
-      <Meta
-        title="Highlight: Plans And Pricing For Any Team. Get Started Free."
-        description="Highlight's developer friendly pricing makes sure any team can afford to get the visibility into bugs they need. See plans, features, FAQs and more here:"
-      />
       <Navbar />
-      <main>
-        <Section className={styles.titleSection}>
-          <div className={styles.pricingTitle}>
-            <div className={styles.anchorHead}>
-              <h1>
-                Get the
-                <span className={styles.yellowText}>{` visibility `}</span>
-                you need today.
-              </h1>
-            </div>
-          </div>
-        </Section>
+      <div className="flex flex-col w-full px-10 mx-auto mt-24">
+        <div className="flex flex-col items-center text-center gap-9">
+          {' '}
+          {/* Title, tagline and CTA */}
+          <h1 className="max-w-3xl">
+            Get the <span className="text-highlight-yellow">visibility</span> you need today.
+          </h1>
+          <Typography type="copyHeader" onDark>
+            Fair and transparent pricing that scales with any organization.
+          </Typography>
+        </div>
+        <PlanTable />
+        <div className="flex flex-col items-center mt-32 text-center gap-9" id="overage">
+          {' '}
+          {/* Pay as you go */}
+          <h2>
+            Pay <span className="text-highlight-yellow">as you go.</span>
+          </h2>
+          <Typography type="copy1" onDark className="max-w-4xl">
+            Each of our plans comes with a pre-defined usage quota, and if you exceed that quota, we charge an
+            additional fee. For custom plans, <a href="#">reach out to us</a>.
+          </Typography>
+        </div>
+        <PriceCalculator />
+      </div>
 
-        {isMobile ? (
-          <Section className={styles.mobileTierWrapper}>
-            <div className={styles.mobileBillingWrapper}>
-              <Typography type="copy3" className={styles.billingCycleText}>
-                Select billing cycle
-              </Typography>
-              <BillingWidget onMonthlyChange={(m) => setMonthly(m)} monthly={monthly} />
-            </div>
-            <MobileTierCarousel monthly={monthly} />
-          </Section>
-        ) : (
-          <Section className={styles.tierWrapper}>
-            <div className={styles.configColumn}>
-              <BillingWidget onMonthlyChange={(m) => setMonthly(m)} monthly={monthly} />
-              <div className={styles.featureKeys}>
-                {Object.keys(BasicDetails).map((headingKey) => {
-                  return (
-                    <div key={headingKey} className={styles.featureSet}>
-                      <Typography className={styles.featureName} type="copy3" emphasis>
-                        {(BasicDetails as any)[headingKey].name}
-                      </Typography>
-                      {(BasicDetails as any)[headingKey].items.map((item: any, key: number) => {
-                        return (
-                          <>
-                            <Typography type="copy3" key={key} className={styles.featureKey} onDark>
-                              {item.name}
-                            </Typography>
-                            <hr className={styles.featureKeyDivider} />
-                          </>
-                        )
-                      })}
-                    </div>
-                  )
-                })}
-              </div>
-            </div>
-            {planInfo.map((p: PricingInfo, idx: number) => (
-              <TierSection key={idx} features={planDetails[idx]} monthly={monthly} {...p} />
-            ))}
-          </Section>
-        )}
-        <Section>
-          <CompaniesReel />
-        </Section>
-        <Section className={styles.faqSection}>
-          <div className={styles.anchorFeature}>
-            <div className={styles.anchorHead}>
-              <h2>{`Frequently asked questions`}</h2>
-            </div>
+      <div className="px-10 mt-32 mb-20">
+        <CompaniesReel />
+
+        <div className={classNames(homeStyles.anchorFeature, 'mt-20')}>
+          <div className={homeStyles.anchorHead}>
+            <h2>{`Frequently asked questions`}</h2>
           </div>
+        </div>
+        <Section>
           <div>
-            {FAQS.map((faq, index) => (
+            {Faqs.map((faq, index) => (
               <Question key={index} question={faq.question} answer={faq.answer} icon={faq.icon} />
             ))}
           </div>
         </Section>
-        <FooterCallToAction />
-      </main>
+      </div>
+      <FooterCallToAction />
       <Footer />
     </div>
   )
 }
 
-const DownArrow = ({ fill }: { fill: string }) => (
-  <svg width={12} height={12} fill="none" xmlns="http://www.w3.org/2000/svg">
-    <g clipPath="url(#a)">
-      <path d="M11.625 3.375 6 9 .375 3.375" stroke={fill} strokeWidth={2} strokeMiterlimit={10} />
-    </g>
-    <defs>
-      <clipPath id="a">
-        <path fill={fill} d="M0 0h12v12H0z" />
-      </clipPath>
-    </defs>
-  </svg>
-)
-
 const Question = ({ question, answer, icon }: { question: string; answer: string; icon: string }) => {
   const [expanded, setExpanded] = useState(false)
   return (
-    <div className={styles.faqRowClickable}>
+    <div className={pricingStyles.faqRowClickable}>
       <Collapsible
         onOpening={() => setExpanded(true)}
         onClosing={() => setExpanded(false)}
         transitionTime={200}
         trigger={
-          <div className={styles.faqRow}>
-            <div className={styles.faqLeftContent}>
+          <div className={pricingStyles.faqRow}>
+            <div className={pricingStyles.faqLeftContent}>
               <Image src={icon} alt="pc icon"></Image>
-              <Typography className={styles.question} type="copy2" emphasis>
+              <Typography className={pricingStyles.question} type="copy2" emphasis>
                 {question}
               </Typography>
             </div>
             <button
-              className={classNames(styles.circleButton, {
-                [styles.expanded]: expanded,
+              className={classNames(pricingStyles.circleButton, {
+                [pricingStyles.expanded]: expanded,
               })}
             >
-              <DownArrow fill={expanded ? '#0D0225' : '#72E4FC'} />
+              <ChevronDownIcon className={classNames('w-5 h-5', expanded ? 'text-dark-background' : 'text-blue-cta')} />
             </button>
           </div>
         }
       >
-        <div className={styles.collapseInner}>
-          <Typography className={styles.questionDescription} type="copy3" onDark>
+        <div className={pricingStyles.collapseInner}>
+          <Typography className={pricingStyles.questionDescription} type="copy3" onDark>
             <div dangerouslySetInnerHTML={{ __html: answer }}></div>
           </Typography>
         </div>
       </Collapsible>
-      <hr className={styles.faqDivider} />
+      <hr className={pricingStyles.faqDivider} />
     </div>
   )
 }
 
-const BillingWidget = ({
-  monthly,
-  onMonthlyChange,
-}: {
-  monthly: boolean
-  onMonthlyChange: (monthly: boolean) => void
-}) => {
-  return (
-    <>
-      <div className={styles.billingWidget}>
-        <div className={styles.billingWidgetButtons}>
-          <PrimaryButton
-            className={classNames(styles.billingButton, styles.leftButton, {
-              [styles.selected]: monthly,
-            })}
-            onClick={() => onMonthlyChange(true)}
-          >
-            Monthly
-          </PrimaryButton>
-          <PrimaryButton
-            className={classNames(styles.billingButton, styles.rightButton, {
-              [styles.selected]: !monthly,
-            })}
-            onClick={() => onMonthlyChange(false)}
-          >
-            Annual
-          </PrimaryButton>
-        </div>
-      </div>
-    </>
-  )
+const docsUrl = '/docs'
+
+const Faqs: { question: string; answer: string; icon: string }[] = [
+  {
+    question: 'Do you offer a discount for non-profits?',
+    answer: `We love supporting non-profits and offer a 75% discount for the lifetime of the account. To activate the discount, create a workplace on either the Standard or Pro plan. Then reach out to support and mention the discount.`,
+    icon: TagLoyalty,
+  },
+  {
+    question: 'How long does it take to setup Highlight?',
+    answer: `It generally takes an engineer less than ten minutes to understand the concepts of Highlight and integrate the app into their workflow. For more information on setup, take a look at our <a href="${docsUrl}">docs</a>.`,
+    icon: Stopwatch,
+  },
+  {
+    question: 'Can I deploy Highlight on-premise?',
+    answer: `Yes! To get a glimpse at how our deployment process looks, take a look <a href="${docsUrl}/on-premise">here</a> (its super simple!). We also support deployments for most cloud providers (Heroku, Render, AWS, etc..). To get a license key for a trial or a production deployment, contact <a href="mailto:sales@highlight.io">sales</a>.`,
+    icon: Globe,
+  },
+  {
+    question: "Is Highlight secure? Where's my data stored?",
+    answer: `Highlight uses end-to-end encryption to keep your data safe while it’s in transit, and we also offer an on-prem solution if you want to keep customer data on your own servers. For more information, see our <a href="/#privacy">security section</a> and <a href="${docsUrl}" target="_blank">docs</a>. If we don't answer your question there, <a href="mailto:jay@highlight.io">let us know</a>.`,
+    icon: Security,
+  },
+  {
+    question: 'Do I need a credit card to sign up?',
+    answer: `Absolutely not! We never ask for your credit card on sign up. If you start on a paid plan then 30 days after signing up you will be politely prompted to enter in your payment information. At anytime you can switch back to a free plan as long as your workplace has less than 6 seats.`,
+    icon: CreditCard,
+  },
+  {
+    question: 'How will you charge me?',
+    answer: `We ask for a credit card. Your credit card information will never touch our servers as we use <a href="https://stripe.com/" target="_blank">Stripe</a> as our payments processor. For Enterprise customers we can do ACH and custom invoices if requested.`,
+    icon: Wallet,
+  },
+  {
+    question: 'How does billing work?',
+    answer: `We charge by usage; or number of sessions collected per month. Our billing system uses prorated billing, meaning you only pay for what you use below each of our thresholds (see above). For example if you move to the Startup plan from the Basic plan in the middle of the month, then you will only be charged for the time you are on the paid plan.`,
+    icon: ReceiptList,
+  },
+  {
+    question: 'What counts as a session?',
+    answer: `A session is contiguous instance of a user's presence on your app for less than 4 hours. That is, if a user is browsing your application for 3 minutes, then closes the tab, this counts as a single session.`,
+    icon: PcPlayMedia,
+  },
+  {
+    question: 'Can I cancel at anytime?',
+    answer: `Definitely! You can cancel or downgrade your subscription at anytime. You can also delete your workplace in the settings page at anytime.`,
+    icon: Delete,
+  },
+]
+
+const billingPeriodOptions = ['Monthly', 'Annual'] as const
+type BillingPeriod = (typeof billingPeriodOptions)[number]
+
+const retentionOptions = ['3 months', '6 months', '1 year', '2 years'] as const
+type Retention = (typeof retentionOptions)[number]
+const retentionMultipliers: Record<Retention, number> = {
+  '3 months': 1,
+  '6 months': 1.5,
+  '1 year': 2,
+  '2 years': 2.5,
+} as const
+
+const tierOptions = ['Free', 'Basic', 'Essentials', 'Startup'] as const
+type TierName = (typeof tierOptions)[number]
+
+type PricingTier = {
+  basePrice: number
+  sessions: number
+  errors: number
 }
 
-const MobileTierCarousel = ({ monthly }: { monthly: boolean }) => {
-  const { width } = useWindowDimensions()
-  const [planIndex, setPlanIndex] = useState(2)
-  const [viewportRef, embla] = useEmblaCarousel({
-    startIndex: planIndex,
-    align: 'center',
-  })
+const priceTiers: Record<TierName, PricingTier> = {
+  Free: { basePrice: 0, sessions: 500, errors: 1000 },
+  Basic: { basePrice: 50, sessions: 2000, errors: 4000 },
+  Essentials: { basePrice: 150, sessions: 10000, errors: 20000 },
+  Startup: { basePrice: 400, sessions: 80000, errors: 160000 },
+}
 
-  useEffect(() => {
-    if (embla && planIndex) {
-      embla.scrollTo(planIndex)
+function getBasePrice({ basePrice }: PricingTier, billing: BillingPeriod, retention: Retention) {
+  const billingMultiplier = billing === 'Annual' ? 0.8 : 1
+  const retentionMultiplier = retentionMultipliers[retention]
 
-      embla.on('select', () => {
-        setPlanIndex(embla.selectedScrollSnap())
-      })
-    }
-  }, [embla, planIndex])
+  return basePrice * billingMultiplier * retentionMultiplier
+}
+
+const PlanTable = () => {
+  const [billingPeriod, setBillingPeriod] = useState<BillingPeriod>('Monthly')
+  const [retention, setRetention] = useState<Retention>('3 months')
 
   return (
-    <>
-      <div className="embla" ref={viewportRef}>
-        <div className="embla__container">
-          {planInfo.map((p: PricingInfo, i: number) => (
-            <MobileTierSection
-              key={i}
-              width={width <= 800 ? 248 : 332}
-              selected={i == planIndex}
-              monthly={monthly}
-              {...p}
-            />
-          ))}
-        </div>
+    <div className="flex flex-col items-center max-w-full gap-6 mx-auto mt-16">
+      {' '}
+      {/* Pricing */}
+      <div className="flex flex-wrap justify-center gap-12 gap-y-3">
+        <RadioOptions
+          title="Billing Period"
+          options={billingPeriodOptions}
+          value={billingPeriod}
+          onChange={setBillingPeriod}
+        />
+        <RadioOptions title="Retention" options={retentionOptions} value={retention} onChange={setRetention} />
       </div>
-      <div className={styles.mobileTierDots}>
-        {planInfo.map((p: PricingInfo, i: number) => (
-          <button
-            key={i}
-            onClick={() => {
-              embla?.scrollTo(i, false)
-            }}
-            className={classNames(styles.mobileTierDot, {
-              [styles.dotSelected]: i == planIndex,
-            })}
-          ></button>
+      <div className="grid items-stretch w-full grid-cols-1 sm:grid-cols-2 min-[1190px]:grid-flow-col gap-7">
+        {Object.entries(priceTiers).map(([name, tier]) => (
+          <PlanTier name={name} tier={tier} billingPeriod={billingPeriod} key={name} retention={retention} />
         ))}
       </div>
-      <div className={styles.mobileFeaturesWrapper}>
-        <Typography type="copy3" emphasis>
-          {Object.keys(planDetails[planIndex]).map((headingKey) => {
-            let currentFeatureSection = (planDetails[planIndex] as any)[headingKey]
-            return (
-              <div className={styles.mobileFeaturesSection} key={headingKey}>
-                <div className={styles.mobileFeatureHeader}>
-                  <Typography type="copy3" emphasis>
-                    {currentFeatureSection.name}
-                  </Typography>
-                </div>
-                {Object.keys(currentFeatureSection.items).map((featureKey: any, i: number) => {
-                  let currentFeature = currentFeatureSection.items[featureKey]
-                  return (
-                    <div key={i} className={styles.mobileFeatureRow}>
-                      <div className={styles.mobileFeatureInner}>
-                        <Typography type="copy3" className={styles.mobileFeatureName}>
-                          {currentFeature.name}
-                        </Typography>
-                        <Typography type="copy3">
-                          {currentFeature.value == true ? (
-                            <Image src={CheckMark} alt="checkmark"></Image>
-                          ) : (
-                            <div>-</div>
-                          )}
-                        </Typography>
-                      </div>
-                      <hr className={styles.mobileFeatureDivider} />
-                    </div>
-                  )
-                })}
-              </div>
-            )
-          })}
-        </Typography>
-      </div>
-    </>
+      <Typography type="copy1" onDark className="text-center my-9">
+        If usage goes beyond the included monthly quota, your <a href="#overage">usage rate</a> kicks in.
+      </Typography>
+      <div className="flex-shrink w-48" />
+    </div>
   )
 }
 
-const MobileTierSection = ({
-  tierName,
-  numSessionCredits,
-  price,
-  mostPopular,
-  selected,
-  width,
-  discount,
-  monthly,
+const PlanTier = ({
+  name,
+  tier,
+  billingPeriod,
+  retention,
 }: {
-  monthly: boolean
-  selected: boolean
-  width: number
-} & PricingInfo) => {
-  const contactSales = price === undefined
+  name: string
+  tier: PricingTier
+  billingPeriod: BillingPeriod
+  retention: Retention
+}) => {
+  const { sessions, errors } = tier
+
   return (
-    <div
-      style={{ minWidth: width }}
-      className={classNames(styles.mobileTierSectionWrapper, 'embla__slide', {
-        [styles.mobileNotSelectedPlan]: !selected,
-      })}
-    >
-      {mostPopular && (
-        <div className={styles.mostPopularMobile}>
-          <Typography type="outline">Most Popular</Typography>
-        </div>
-      )}
-      <div
-        className={classNames(styles.mobileTier, {
-          [styles.mostPopularTierColumnMobile]: mostPopular,
-        })}
-      >
-        <div className={styles.mobileTierTop}>
-          <div className={styles.tierHeader}>
-            <Typography type="copy1" emphasis>
-              {tierName}
-            </Typography>
-            <div
-              className={classNames(styles.discountPill, {
-                [styles.discountPillVisible]: !monthly && discount,
-              })}
-            >
-              - {discount}%
-            </div>
-          </div>
-          <Typography type="copy3" className={styles.mobileSessionCredits}>
-            {numSessionCredits ? numSessionCredits.toLocaleString() : 'Custom'} session credits
+    <div className="flex flex-col flex-grow border rounded-md min-[1190px]:min-w-[255px] basis-64 border-divider-on-dark">
+      <div className="p-5 border-b border-divider-on-dark">
+        <Typography type="copy1" emphasis>
+          {name}
+        </Typography>
+        <div className="flex items-end mt-2">
+          <Typography type="copy3" emphasis className="self-start align-super">
+            $
           </Typography>
-          <div className={styles.desktopPrice}>
-            {contactSales ? (
-              <Image height={24} width={24} src={Chat} alt="chat icon"></Image>
-            ) : (
-              <>
-                <Typography type="copy3" emphasis className={styles.moneySign}>
-                  {'$'}
-                </Typography>
-                <h3 className={styles.price}>
-                  {monthly || !discount ? price : Math.round((1 - discount / 100) * price)}
-                </h3>
-                <div className={styles.timeIndicator}>
-                  <Typography type="copy3">{'/ mo'}</Typography>
-                </div>
-              </>
-            )}
-          </div>
+          <span className="mx-1 text-5xl font-semibold">{getBasePrice(tier, billingPeriod, retention)}</span>
+          <Typography type="copy3">/ mo</Typography>
         </div>
-        <PrimaryButton className={classNames(styles.pricingButton, styles.mobilePricingButton)}>
-          <Typography type="copy2" emphasis={true}>
-            {contactSales ? 'Contact Sales' : 'Start Free Trial'}
-          </Typography>
+      </div>
+      <div className="p-5 flex flex-col gap-2.5 flex-grow">
+        <div className="flex items-center gap-1">
+          <Typography type="copy3" emphasis>
+            Included
+          </Typography>{' '}
+          <InformationCircleIcon className="inline w-5 h-5" />
+        </div>
+        <Typography type="copy3">{sessions} monthly sessions</Typography>
+        <Typography type="copy3">{errors} monthly errors</Typography>
+        <Typography type="copy3">Unlimited seats</Typography>
+      </div>
+      <div className="p-5">
+        <PrimaryButton href="#" className={homeStyles.hollowButton}>
+          Start free trial
         </PrimaryButton>
       </div>
     </div>
   )
 }
 
-function getWindowDimensions() {
-  const { innerWidth: width, innerHeight: height } = window
-  return {
-    width,
-    height,
-  }
+const formatPrice = (price: number) =>
+  price.toLocaleString(undefined, { style: 'currency', currency: 'USD', signDisplay: 'always' }).replace('+', '+ ')
+
+const PriceCalculator = () => {
+  const [billingPeriod, setBillingPeriod] = useState<BillingPeriod>('Monthly')
+  const [tierName, setTierName] = useState<TierName>('Free')
+  const [retention, setRetention] = useState<Retention>('3 months')
+
+  const [errorUsage, setErrorUsage] = useState(0)
+  const [sessionUsage, setSessionUsage] = useState(0)
+  const [loggingUsage, setLoggingUsage] = useState(0)
+
+  const tier = priceTiers[tierName]
+  const basePrice = getBasePrice(tier, billingPeriod, retention)
+
+  const getUsagePrice = (useage: number, price: number, size: number) =>
+    Math.trunc(((Math.max(useage, 0) * price) / size) * retentionMultipliers[retention] * 100) / 100
+
+  const sessionsCost = getUsagePrice(sessionUsage - tier.sessions, 5.0, 1_000)
+  const errorsCost = getUsagePrice(errorUsage - tier.errors, 0.2, 1_000)
+  const loggingCost = getUsagePrice(loggingUsage, 1.5, 1_000_000)
+
+  return (
+    <div className="flex flex-col items-center w-full gap-10 mx-auto mt-12">
+      {' '}
+      {/* Price calculator */}
+      <div className="flex flex-wrap justify-center gap-12 gap-y-3">
+        <RadioOptions
+          title="Billing Period"
+          options={billingPeriodOptions}
+          value={billingPeriod}
+          onChange={setBillingPeriod}
+        />
+        <RadioOptions title="Pricing Tier" options={tierOptions} value={tierName} onChange={setTierName} />
+        <RadioOptions title="Retention" options={retentionOptions} value={retention} onChange={setRetention} />
+      </div>
+      <div className="flex flex-col items-end w-full max-w-[1100px]">
+        <div className="flex flex-col overflow-hidden border divide-y rounded-lg md:rounded-br-none divide-divider-on-dark border-divider-on-dark">
+          <div className="hidden h-12 md:flex">
+            <div className="flex items-center flex-1 border-r border-divider-on-dark px-7">
+              <Typography type="copy2" emphasis>
+                Product
+              </Typography>
+            </div>
+            <div className="flex items-center justify-center w-[343px] px-7">
+              <Typography type="copy2" emphasis>
+                Cost breakdown
+              </Typography>
+            </div>
+          </div>
+          <CalculatorRowDesktop
+            title="Error Monitoring Usage."
+            description="Error monitoring usage is defined by the number of errors collected by Highlight per month. Our frontend/server SDKs send errors, but you can also send custom errors."
+            value={errorUsage}
+            onChange={setErrorUsage}
+            cost={errorsCost + basePrice}
+          />
+          <CalculatorRowDesktop
+            title="Session Replay Usage."
+            description="Session replay usage is defined by the number of sessions collected per month. A session is defined by an instance of a user’s tab on your application. "
+            value={sessionUsage}
+            onChange={setSessionUsage}
+            cost={sessionsCost + basePrice}
+          />
+          <CalculatorRowDesktop
+            title="Logging Usage."
+            description="Log usage is defined by the number of logs collected by highlight.io per month. A log is defined by a text field with attributes."
+            value={loggingUsage}
+            onChange={setLoggingUsage}
+            cost={loggingCost + basePrice}
+          />
+          <div className="block px-3 py-5 rounded-b-lg md:hidden">
+            <Typography type="copy1" emphasis>
+              Total: {formatPrice(basePrice + sessionsCost + errorsCost + loggingCost)}
+            </Typography>
+          </div>
+        </div>
+        <div className="hidden border border-t-0 rounded-b-lg md:block h-52 border-divider-on-dark">
+          <CalculatorCostDisplay heading="Total" cost={basePrice + sessionsCost + errorsCost + loggingCost} />
+        </div>
+      </div>
+    </div>
+  )
 }
 
-function useWindowDimensions() {
-  const [windowDimensions, setWindowDimensions] = useState(getWindowDimensions())
+const CalculatorRowDesktop = ({
+  title,
+  description,
+  value,
+  onChange,
+  cost,
+}: {
+  title: string
+  description: string
+  value: number
+  onChange: (value: number) => void
+  cost: number
+}) => {
+  const rangeOptions = [0, 500, 1_000, 5_000, 10_000, 100_000, 250_000, 500_000, 750_000, 1_000_000]
 
-  useEffect(() => {
-    function handleResize() {
-      setWindowDimensions(getWindowDimensions())
+  return (
+    <div className="flex flex-row">
+      <div className="flex flex-col flex-1 gap-1 px-3 py-5 md:px-7">
+        <Typography
+          type="copy4"
+          emphasis
+          className="rounded-full bg-blue-cta py-0.5 px-3 text-dark-background self-start inline-block md:hidden"
+        >
+          {formatPrice(cost)}
+        </Typography>
+        <Typography type="copy1" emphasis>
+          {title}
+        </Typography>
+        <Typography type="copy3" className="mt-2.5">
+          {description}
+        </Typography>
+        <RangedInput options={rangeOptions} value={value} onChange={onChange} />
+      </div>
+      <div className="hidden border-l border-divider-on-dark md:inline-block">
+        <CalculatorCostDisplay heading="Base + Usage" cost={cost} />
+      </div>
+    </div>
+  )
+}
+
+export const RangedInput = ({
+  options,
+  value,
+  onChange,
+}: {
+  options: number[]
+  value: number
+  onChange: (value: number) => void
+}) => {
+  const sortedOptions = [...options].sort((a, b) => a - b)
+  const min = sortedOptions[0] ?? 0
+  const max = sortedOptions[sortedOptions.length - 1] ?? 100
+
+  const snapValue = (value: number) => {
+    const deltas = sortedOptions.map((v) => Math.abs(v - value))
+    for (const [i, delta] of deltas.entries()) {
+      if ((deltas[i + 1] ?? Infinity) > delta) return sortedOptions[i]
     }
+    return sortedOptions[0]
+  }
 
-    window.addEventListener('resize', handleResize)
-    return () => window.removeEventListener('resize', handleResize)
-  }, [])
+  const normalize = (value: number) => (value - min) / (max - min)
+  const denormalize = (normal: number) => normal * (max - min) + min
 
-  return windowDimensions
+  return (
+    <>
+      <div className="block md:hidden">
+        <div className="relative">
+          <select
+            className="flex items-center justify-center w-full h-12 gap-2 text-center text-transparent transition-all border rounded-lg appearance-none cursor-pointer border-copy-on-light hover:bg-white/10 bg-dark-background"
+            onChange={(ev) => onChange(parseFloat(ev.target.value))}
+          >
+            {options.map((value, i) => (
+              <option value={value} key={i}>
+                {value.toLocaleString(undefined, { notation: 'compact' })}
+              </option>
+            ))}
+          </select>
+          <div className="absolute inset-0 flex items-center justify-center w-full h-12 gap-2 transition-all border rounded-lg pointer-events-none border-copy-on-light hover:bg-white/10">
+            <Typography type="copy2" emphasis onDark>
+              {snapValue(value).toLocaleString(undefined, { notation: 'compact' })}
+            </Typography>
+            <ChevronDownIcon className="w-5 h-5 text-darker-copy-on-dark" />
+          </div>
+        </div>
+      </div>
+      <Slider.Root
+        min={min}
+        max={max}
+        step={500}
+        value={[denormalize(Math.pow(normalize(value), 1 / 3))]}
+        onValueChange={([value]) => value != null && onChange(Math.ceil(denormalize(Math.pow(normalize(value), 3))))}
+        className="relative items-center hidden w-full h-16 mt-4 select-none md:flex touch-none group"
+      >
+        <Slider.Track className="relative flex-1 h-3 overflow-hidden rounded-full bg-divider-on-dark" />
+        <Slider.Thumb className="relative w-6 h-6 border-2 focus:border-purple-primary hover:shadow-white/25 hover:shadow-[0_0_0_4px] outline-none bg-[#F5F5F5] border-copy-on-dark rounded-full flex flex-col items-center transition-all">
+          <div className="absolute w-2.5 h-2.5 rotate-45 rounded-sm -top-4 bg-blue-cta" />
+          <div className="absolute px-1 py-0.5 mb-2 text-divider-on-dark font-semibold text-[10px] rounded-sm bottom-full bg-blue-cta">
+            {value.toLocaleString(undefined, { notation: 'compact' })}
+          </div>
+        </Slider.Thumb>
+      </Slider.Root>
+    </>
+  )
 }
 
-export default Pricing
+const CalculatorCostDisplay = ({ cost, heading }: { cost: number; heading: string }) => (
+  <div className="grid flex-shrink-0 place-content-center place-items-center w-[343px] h-full">
+    <Typography type="copy3" emphasis onDark>
+      {heading}
+    </Typography>
+    <span className="text-4xl font-semibold">{formatPrice(cost)}</span>
+  </div>
+)
+
+const RadioOptions = <T extends string>({
+  title,
+  options,
+  value,
+  onChange,
+}: {
+  title: string
+  options: readonly T[]
+  value?: T
+  onChange?: (value: T) => void
+}) => {
+  return (
+    <RadioGroup value={value} onChange={onChange} className="flex flex-col items-center gap-2">
+      <RadioGroup.Label className="">
+        <Typography type="copy4" className="text-center text-darker-copy-on-dark">
+          {title}
+        </Typography>
+      </RadioGroup.Label>
+      <div className="flex p-px border rounded-[10px] gap-1 bg-user-black border-divider-on-dark">
+        {options.map((option) => (
+          <RadioGroup.Option value={option} key={option}>
+            {({ checked }) => (
+              <div className="cursor-pointer">
+                <div
+                  className={classNames(
+                    'text-center px-2.5 py-1.5 select-none rounded-lg transition-colors',
+                    checked ? 'text-dark-background bg-white' : 'text-white hover:bg-white/10',
+                  )}
+                >
+                  <Typography type="copy3" emphasis>
+                    {option}
+                  </Typography>
+                </div>
+              </div>
+            )}
+          </RadioGroup.Option>
+        ))}
+      </div>
+    </RadioGroup>
+  )
+}
+
+export default PricingPage
